redux: refactored getProductFilter -> getAllProductsFiltered, action-type GET_PRODUCT_FILTER -> GET_ALL_PRODUCTS_FILTERED
