+++ src/redux/action-type.js
@@ -22,7 +22,7 @@ export const SET_USER = 'SET_USER'
 export const GET_USER_BY_ID = 'GET_USER_BY_ID'
 
 // FILTERS
-export const GET_PRODUCT_FILTER = 'GET_PRODUCT_FILTER'
+export const GET_ALL_PRODUCTS_FILTERED = 'GET_ALL_PRODUCTS_FILTERED'
 export const GET_BEST_SELL = 'GET_BEST_SELL'
 export const SET_CATEGORY = 'SET_CATEGORY'
 export const SET_HIGH_PRICE = 'SET_HIGH_PRICE'
